Moved SchemaViewer to folder
