+++ packages/eventcatalog/components/SchemaViewer/SchemaViewer.module.css
