+++ packages/eventcatalog/components/SchemaViewer/SchemaViewer.tsx
